+++ x/bank/migrations/v4/store.go
@@ -21,6 +21,9 @@ func MigrateStore(ctx sdk.Context, storeService store.KVStoreService, legacySubs
 	var currParams types.Params
 	legacySubspace.GetParamSet(ctx, &currParams)
 
+	// SendEnabled is migrated to the x/bank module store, so delete from the params
+	currParams = types.NewParams(currParams.DefaultSendEnabled)
+
 	if err := currParams.Validate(); err != nil {
 		return err
 	}